+++ src/aui-form-builder/HISTORY.md
@@ -4,6 +4,7 @@
 
 ## @VERSION@
 
+* [AUI-2032](https://issues.liferay.com/browse/AUI-2032) Move out logic that displays field types modal
 * [AUI-2016](https://issues.liferay.com/browse/AUI-2016) Remove column delete button from From Builder
 * [AUI-2019](https://issues.liferay.com/browse/AUI-2019) The Active page number shouldn't appear over the Fields Panel
 * [AUI-2018](https://issues.liferay.com/browse/AUI-2018) Create a template variable to form builder modal head content